Create test_vdocs.py Update test_emulators.py
